+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Generator configuration by build property
+- Added EmbeddingFileNameType config
 
 ## [4.1.1] - 2022-02-26
 ### Changed